CtrlLib: ArrayCtrl::GetCellRect[M] now public
git-svn-id: svn://ultimatepp.org/upp/trunk@10916 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlLib/Progress.cpp
@@ -57,7 +57,7 @@ void ProgressIndicator::Paint(Draw& w) {
 		w.DrawRect(r2, SColorPaper);
 		w.DrawRect(r3, SColorPaper);
 		if(percent) {
-			String pt = Format("%d %%", 100 * actual / max(total, 1));
+			String pt = Format("%d %%", (int)(100L * actual / max(total, 1)));
 			Size psz = GetTextSize(pt, StdFont());
 			int px = (sz.cx - psz.cx) / 2 + 2;
 			int py = (sz.cy - psz.cy) / 2 + 2;
